Add top-level functional test
The subject of this test is the composite middleware made of callable middlewares. Considering other interfaces in this package, it would make most sense to inject the upcoming SUT implementation with other factory implementations, and to return nested middleware and pipe implementations. In such a scenario, this test will test most of the functionality in this package in one go. This is not very granular, but is a simple and structured way to start working on the package, and an easy way to smoke-test this.